docs: MainPage, MainComponents로 분리
