+++ src/Pages/MainPage/styled.ts
@@ -10,4 +10,10 @@ export const MapBox = styled.div`
   justify-content: center;
   align-items: center;
 `
+export const MapContainer = styled.div`
+  width: 100%;
+  height: 100%;
+`;
+
+
 export const Box = styled.div``
